Refactoring ArtistInfoActivity
Muevo el archivo al la carpeta moduel
Creo interfaz asociada
Correción en OtherInfoActivity
